refactor: move general generate functions to utils
